requirements.txt
alembic==1.6.5
attrs==21.2.0
bleach==3.3.0
click==8.0.1
Flask==2.0.1
Flask-Login==0.5.0
Flask-WTF==0.15.1
greenlet==1.1.0
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.1
Mako==1.1.4
MarkupSafe==2.0.1
marshmallow==3.12.1
packaging==20.9
pluggy==0.13.1
psycopg2==2.9.1
py==1.10.0
pyparsing==2.4.7
pytest==6.2.4
python-dateutil==2.8.1
python-dotenv==0.18.0
python-editor==1.0.4
six==1.16.0
SQLAlchemy==1.4.19
toml==0.10.2
webencodings==0.5.1
Werkzeug==2.0.1
WTForms==2.3.3
